feat: Make loading from an external URL be gated behind a feature. (#149)
* feat: Make loading from an external URL be gated behind a feature.

`attohttpc` and its transitive dependencies add a large amount of size
to the crate (on the order of about ~400 kB). For applications which
don't need external URL loading and are concerned with code size (for
example, a WASM binary), this is wasted space. Placing this stuff behind
a feature gate allows people to opt-out and reduce code size if they wish.

* fix: Add `http` feature by default for python and wasm bindings.

I'm assumining that these bindings expect `http` to be active. Having
`default_features` be false meant that these bindings weren't getting
the `http` gated code.